refactor: update tensor backend to support dtype-native operations and improve documentation
- Refactored tensor backend to transition from Float64Tensor to a more generic Tensor type, enhancing compatibility across various data types.
- Updated README.md to reflect changes in build commands and backend structure, specifically for CUDA and Metal devices.
- Enhanced VERIFICATION_STATUS.md with detailed benchmark outputs and test results for the Metal device backend.
- Introduced new helper functions for tensor value extraction and conversion, improving code clarity and usability.
- Updated tests across CPU and CUDA backends to align with the new tensor structure, ensuring consistent behavior and performance.
- Improved error handling and documentation throughout the backend implementation, providing clearer guidance for developers.

+++ pkg/hub/client_test.go
@@ -331,7 +331,7 @@ func TestClient_DownloadXet(test *testing.T) {
 			case "/api/models/org/repo/xet-read-token/main":
 				writeJSON(writer, map[string]any{
 					"accessToken": "xet_read",
-					"exp":         9999999999,
+					"exp":         int64(9999999999),
 					"casUrl":      serverURL + "/cas",
 				})
 			case "/cas/v1/reconstructions/" + fileID:
